Remove the direct numpy dependency
It's still an indirect dependency through the h5py dev requirement. But
that's only needed for generating a quantized model, not for general
usage.

The motivation for this is that tests are failing on Windows, in a
manner similar to what's described in the comments here:
https://github.com/pytorch/pytorch/issues/128860

The error suggests that the torch wheel for Windows was generated with
numpy<2.